requirements.txt
certifi>=2024.0.0
feedparser>=6.0.11
httpx>=0.27.0
sentence-transformers>=3.3.0
numpy>=1.26.0
PyYAML>=6.0
torch>=2.2.0
transformers>=4.40.0
